Maximum Liklihood Estimate with Finite Sample Approximation Indicator

Function to get MLE estimates for data with the related cutoff based on finite sample approximation.

Usage

MLEVostIndicator(
  u,
  y,
  lower,
  upper,
  alpha = 0.05,
  nStart = NA,
  nEnd = NA,
  returnEstims = FALSE,
  estimVarLS = FALSE,
  var1BaseEstim = 0.5,
  var2BaseEstim = 0.5,
  maxOptimIters = 2
)

Arguments

u

Vector of numerics for the 4 parameters of interest, c(b1,s1^2,s2^2,b2). This is an initial estimate

y

Vector of numerics for the data

lower

Vector of numerics indicating the lower bound.Variance cannot be negative and the third term must also be above 0.

upper

Vector of numerics indicating the upper bound.

alpha

(Optional) Numeric indicating the significance of interest. Default is 0.05.

nStart

(Optional) Integer indicating starting point to check data. Default results in minimal trimming (4 for params).

nEnd

(Optional) Integer indicating ending point to check data. Default results in minimal trimming (4 for params).

returnEstims

(Optional) Boolean indicating if the estimated parameters should be returned. Default is FALSE.

estimVarLS

(Optional) Boolean to indicate if we should use LS to estimate var2. Default is FALSE.

var1BaseEstim

(Optional) Numeric indicating base var1 estimate. Default is 0.5.

var2BaseEstim

(Optional) Numeric indicating base var1 estimate. Default is 0.5. May not be used if estimVarLS.

maxOptimIters

(Optional) Numeric indicating number of times to re-try optimization. Can sometimes add if get randomly odd result. Default is 2.

Value

Depends on returnEstims

If true: List with element 1 being the WLS estimates, element 2 being the MLE estimates, and element 3 being vector with sqrt Tn estimate, cutoff, location of max Tn, and LNN.

If false: Vector with sqrt Tn estimate, cutoff, and loc of max Tn.

Examples

data <- generateRCA1Data(pars=c(0.5,1,0.25,-0.5), k=250, burnin=1000,
                        iterations=500)
MLEVostIndicator(c(0.4,0.9,0.1,-0.1), data, c(-Inf,0,10^-8,-Inf),
             c(Inf,Inf,Inf,Inf))
#> [1]   7.084637   3.261324 254.000000