Binary Segmentation for Change Point Detection

This function uses binary segmentation to detect change points.

Usage

binarySegmentationCPDetection(
  fullData,
  method,
  lower,
  upper,
  alpha,
  estimVarLS = FALSE,
  var1BaseEstim = 0.5,
  var2BaseEstim = NA,
  maxOptimIters = 1,
  trimAmt = 4,
  verifyCPs = TRUE,
  silent = FALSE
)

Arguments

fullData

Data.frame with two columns. The first has date and the second has values

method

String indicating method to use for change point detection. Options are 'MLE', 'Vostrikova' , and 'WLS'.

lower

Vector of numerics indicating the lower bound. Variance cannot be negative and the third term must also be above 0.

upper

Vector of numerics indicating the upper bound.

alpha

Numeric indicating the significance of interest.

estimVarLS

(Optional) Boolean to indicate if we should use LS to estimate var2. Default is FALSE.

var1BaseEstim

(Optional) Numeric indicating base var1 estimate. Default is 0.5.

var2BaseEstim

(Optional) Numeric indicating base var1 estimate. Default is NA. May not be used if estimVarLS.

maxOptimIters

(Optional) Numeric indicating number of times to re-try optimization. Can sometimes add if get randomly odd result. Default is 1.

trimAmt

(Optional) Number indicating the amount to trim. Default is 4.

verifyCPs

(Optional) Boolean indicating if change points should be rechecked at the end. Default is TRUE.

silent

(Optional) Boolean indicating if output should be supressed. Default is FALSE

Value

Data.frame with columns 'stDate' (start date of region), 'enDate' (end date of region), 'WLS_b1' (beta1 estimate using WLS), 'WLS_v1' (var1 estimate using WLS), 'WLS_v2' (var2 estimate using WLS), 'WLS_b2' (beta2 estimate using WLS), 'MLE_b1' (beta1 estimate using MLE), 'MLE_v1' (var1 estimate using MLE), 'MLE_v2' (var2 estimate using MLE), and 'MLE_b2' (beta2 estimate using MLE).

Examples

data <- generateRCA1Data(pars=c(0.5,1,0.25,-0.5), k=100, burnin=1000,
                        iterations=200)
data <- data.frame(1:length(data),data)
binarySegmentationCPDetection(data,'Vostrikova',
                              lower=c(-Inf,0,10^-8,-Inf),
                              upper=c(Inf,Inf,Inf,Inf), alpha=0.05)
#> 
#> --Find CPs --
#> ChangePoint Detected (1-200 at 99): Segment Data and Re-Search
#> 
#> -- Verify Step --
#> ChangePoint Detected (1-200 at 99): Segment Data and Re-Search
#>   stDate enDate     WLS_b1    WLS_v1    WLS_v2 WLS_b2     MLE_b1    MLE_v1
#> 1      1     99  0.4246383 0.3651574 0.2540698     NA  0.4427789 0.4361462
#> 2    100    200 -0.3678795 0.4208205 0.2901665     NA -0.3690914 0.3614487
#>     MLE_v2 MLE_b2
#> 1 0.253886     NA
#> 2 0.339644     NA