Simulate Power Curve

Simulate data, test for change using 3 methods, plot power curve

Usage

simulatePowerCurve(
  Us3,
  U4s,
  nSims,
  lowerEst,
  upperEst,
  alpha,
  burnin,
  k,
  N,
  errorType = "Normal",
  par1HetereoLocation = NA,
  par1HetereoMult = NA,
  par2HetereoLocation = NA,
  par2HetereoMult = NA,
  trimAmt = 4,
  silent = FALSE
)

Arguments

Us3

Vector of numerics indicating the first 3 parameters

U4s

Vector of numerics indicating the 4th parameters (i.e. multiple to examine)

nSims

Integer indicating the number of simulations for each trial

lowerEst

Vector of numerics indicating the lower bound for the 4 parameters. Variance cannot be negative and the third term must also be above 0.

upperEst

Vector of numerics indicating the upper bound for the 4 parameters.

alpha

Numeric indicating the significance of interest.

burnin

Integer indicating the number of burnin iteration to use

k

Integer indicating the point at when the process should change from using beta1 to using beta2

N

Integer indicating how many observations the final data should be.

errorType

(Optional) String indicating the type of error. Options are 'Normal', 'Bernoulli', and 'Exponential'. Default is Normal.

par1HetereoLocation

(Optional) This can be used to allow hetereoskedasticity in the par1 (2nd term in pars) error. This determines location for it.

par1HetereoMult

(Optional) This can be used to allow hetereoskedasticity in the par1 (2nd term in pars) error. This determines the amount to multiply par1 by for it.

par2HetereoLocation

(Optional) This can be used to allow hetereoskedasticity in the par2 (3rd term in pars) error. This determines location for it.

par2HetereoMult

(Optional) This can be used to allow hetereoskedasticity in the par2 (3rd term in pars) error. This determines the amount to multiply par2 by for it.

trimAmt

(Optional) Numeric indicating amount to trim. Default is 4.

silent

(Optional) Boolean indicating if progress should be output. Default is FALSE

Value

List with two terms.

Plot of power curve

Data.frame of data used for power curve

Examples

pc1 <- simulatePowerCurve(Us3 = c(0, 0.5, 0.5), U4s = c(0,0.5),
                  nSims = 5, lowerEst = c(-Inf,0,10^-8,-Inf),
                  upperEst = rep(Inf,4), alpha = 0.05,
                  burnin = 500, k = 0.5 * 100, N = 100,
                  errorType = 'Normal')
#> Completed: 1/2 - 1/5 (0.107mins remaining)
#> Completed: 1/2 - 2/5 (0.095mins remaining)
#> Completed: 1/2 - 3/5 (0.09mins remaining)
#> Completed: 1/2 - 4/5 (0.075mins remaining)
#> Completed: 1/2 - 5/5 (0.063mins remaining)
#> Completed: 2/2 - 1/5 (0.051mins remaining)
#> Completed: 2/2 - 2/5 (0.039mins remaining)
#> Completed: 2/2 - 3/5 (0.026mins remaining)
#> Completed: 2/2 - 4/5 (0.013mins remaining)
#> Completed: 2/2 - 5/5 (0mins remaining)