Simulation Table with Null Case

This function is used to generate a table of null rejection frequencies

Usage

simulationTable_NC(
  betas,
  varProbRates,
  nSims,
  iterations,
  burnin,
  lowerEstim,
  upperEstim,
  alpha,
  errorTypes,
  CPLoc = 0.5,
  seed = NA,
  silent = F,
  trimAmt = 4
)

Arguments

betas

Vector of numerics indicating betas to examine (remember this is the null case so only one beta used each iteration)

varProbRates

Vector of 2 values indicating the variance or probabilitise used in parameters, c(u2, u3).

nSims

Integer indicating the number of simulations for each trial

iterations

Integer indicating how many observations the final data should be.

burnin

Integer indicating the number of burnin iteration to use

lowerEstim

Vector of numerics indicating the lower bound for the 4 parameters. Variance cannot be negative and the third term must also be above 0.

upperEstim

Vector of numerics indicating the upper bound for the 4 parameters.

alpha

Numeric indicating the significance of interest.

errorTypes

Vector indicating the errorTypes to investigate. Options are 'Normal', 'Bernoulli', and 'Exponential'.

CPLoc

Numeric indicating change point location, between 0 and 1.

seed

(Optional) Integer to indicate seed to set at each interation. Default of NA sets no seed.

silent

(Optional) Boolean indicating if progress should be output. Default is FALSE

trimAmt

(Optional) Number indicating the amount to trim. Default is 4.

Value

Data.frame of null rejection frequencies

Examples

data_er <- simulationTable_NC(betas = c(0.5),
                              varProbRates = c(0.5, 0.5),
                              nSims=5, iterations = c(50),
                              burnin = 1000, lowerEst=c(-Inf,0,10^-8,-Inf),
                              upperEst=c(Inf,Inf,Inf,Inf), alpha = 0.05,
                              errorTypes = c('Normal'),
                              CPLoc=0.5)
#> ErrorType: Normal (1/1)
#> - Starting Beta: 0.5 (1/1)
#> -- Running Iters: 50 (1/1)

if (FALSE) {
data_er <- simulationTable_NC(betas = c(0.5,0.75,1,1.05),
                              varProbRates = c(0.5, 0.5),
                              nSims=500, iterations = c(100, 200, 400, 800),
                              burnin = 1000, lowerEst=c(-Inf,0,10^-8,-Inf),
                              upperEst=c(Inf,Inf,Inf,Inf), alpha = 0.05,
                              errorTypes = c('Normal','Bernoulli','Exponential'),
                              CPLoc=0.5, seed=1234)
}